chore: seperate lib/visualize & plot histograms into one figure
